:recycle: rename helper files format files. return in tests
