+++ src/transactions/utils/index.js
@@ -18,7 +18,7 @@ export {
 export { default as getTransactionBytes } from './getTransactionBytes';
 export { default as getTransactionHash } from './getTransactionHash';
 export { default as getTransactionId } from './getTransactionId';
-export { prependPlusToPublicKeys, prependMinusToPublicKeys } from './helper';
+export { prependPlusToPublicKeys, prependMinusToPublicKeys } from './format';
 export { default as prepareTransaction } from './prepareTransaction';
 export {
 	signTransaction,
